+++ packages/video_player/video_player/test/video_player_test.dart
@@ -73,6 +73,11 @@ class FakeController extends ValueNotifier<VideoPlayerValue>
   Future<void> setLooping(bool looping) async {}
 
   @override
+  Future<void> setPreventsDisplaySleepDuringVideoPlayback(
+    bool prevents,
+  ) async {}
+
+  @override
   VideoFormat? get formatHint => null;
 
   @override
@@ -1858,7 +1863,8 @@ void main() {
           'volume: 0.5, '
           'playbackSpeed: 1.5, '
           'errorDescription: null, '
-          'isCompleted: false),',
+          'isCompleted: false, '
+          'preventsDisplaySleepDuringVideoPlayback: true),',
         );
       });
 
@@ -1952,6 +1958,24 @@ void main() {
 
         await controller.initialize();
         expect(controller.videoPlayerOptions!.mixWithOthers, true);
+      });
+
+      test('setPreventsDisplaySleepDuringVideoPlayback', () async {
+        final controller = VideoPlayerController.networkUrl(_localhostUri);
+        addTearDown(controller.dispose);
+
+        await controller.initialize();
+        expect(controller.value.preventsDisplaySleepDuringVideoPlayback, true);
+
+        await controller.setPreventsDisplaySleepDuringVideoPlayback(false);
+        expect(controller.value.preventsDisplaySleepDuringVideoPlayback, false);
+
+        expect(
+          fakeVideoPlayerPlatform.calls.contains(
+            'setPreventsDisplaySleepDuringVideoPlayback',
+          ),
+          true,
+        );
       });
 
       test('true allowBackgroundPlayback continues playback', () async {
@@ -2225,6 +2249,14 @@ class FakeVideoPlayerPlatform extends VideoPlayerPlatform {
   @override
   Future<void> setMixWithOthers(bool mixWithOthers) async {
     calls.add('setMixWithOthers');
+  }
+
+  @override
+  Future<void> setPreventsDisplaySleepDuringVideoPlayback(
+    int playerId,
+    bool preventsDisplaySleepDuringVideoPlayback,
+  ) async {
+    calls.add('setPreventsDisplaySleepDuringVideoPlayback');
   }
 
   @override