+++ packages/video_player/video_player_avfoundation/lib/src/avfoundation_video_player.dart
@@ -179,6 +179,16 @@ class AVFoundationVideoPlayer extends VideoPlayerPlatform {
   }
 
   @override
+  Future<void> setPreventsDisplaySleepDuringVideoPlayback(
+    int playerId,
+    bool preventsDisplaySleepDuringVideoPlayback,
+  ) {
+    return _playerWith(id: playerId).setPreventsDisplaySleepDuringVideoPlayback(
+      preventsDisplaySleepDuringVideoPlayback,
+    );
+  }
+
+  @override
   Future<List<VideoAudioTrack>> getAudioTracks(int playerId) async {
     final List<MediaSelectionAudioTrackData> nativeData = await _playerWith(
       id: playerId,
@@ -274,6 +284,12 @@ class _PlayerInstance {
   Future<void> setVolume(double volume) => _api.setVolume(volume);
 
   Future<void> setPlaybackSpeed(double speed) => _api.setPlaybackSpeed(speed);
+
+  Future<void> setPreventsDisplaySleepDuringVideoPlayback(
+    bool preventsDisplaySleepDuringVideoPlayback,
+  ) => _api.setPreventsDisplaySleepDuringVideoPlayback(
+    preventsDisplaySleepDuringVideoPlayback,
+  );
 
   Future<void> seekTo(Duration position) {
     return _api.seekTo(position.inMilliseconds);