+++ packages/video_player/video_player_avfoundation/pigeons/messages.dart
@@ -93,4 +93,8 @@ abstract class VideoPlayerInstanceApi {
   List<MediaSelectionAudioTrackData> getAudioTracks();
   @ObjCSelector('selectAudioTrackAtIndex:')
   void selectAudioTrack(int trackIndex);
+  @ObjCSelector('setPreventsDisplaySleepDuringVideoPlayback:')
+  void setPreventsDisplaySleepDuringVideoPlayback(
+    bool preventsDisplaySleepDuringVideoPlayback,
+  );
 }