+++ packages/video_player/video_player_platform_interface/test/video_player_options_test.dart
@@ -14,4 +14,11 @@ void main() {
     final options = VideoPlayerOptions();
     expect(options.mixWithOthers, false);
   });
+  test(
+    'VideoPlayerOptions preventsDisplaySleepDuringVideoPlayback defaults to true',
+    () {
+      final options = VideoPlayerOptions();
+      expect(options.preventsDisplaySleepDuringVideoPlayback, true);
+    },
+  );
 }