+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,5 +1,7 @@
-## NEXT
+## 6.7.0
 
+* Adds `preventsDisplaySleepDuringVideoPlayback` to `VideoPlayerOptions` and
+  `setPreventsDisplaySleepDuringVideoPlayback` to `VideoPlayerPlatform`.
 * Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.
 
 ## 6.6.0